feat: split dsql tree to different chunks
