+++ src/common/monaco/dsql/matchAllQueries.ts
@@ -1,0 +1,12 @@
+export const matchAllQueries = {
+  match_all: {
+    label: 'match_all',
+    snippet: `match_all: {$0},`,
+    children: {
+      boost: {
+        label: 'boost',
+        snippet: 'boost: $1',
+      },
+    },
+  },
+};
